better buyPower moved handleIncomingTransactions make codeclimate  happy
